24/6/19 - 유상 설문 지도 리뷰
전체적 부분 수정및 css 빼고 완성

+++ Anihealth_Project/src/main/java/com/kh/aniht/survey/controller/SurveyController.java
@@ -39,7 +39,7 @@ public class SurveyController {
 	}
 	// 설문결과 출력
 	@PostMapping("surveyResult.su")
-	public String surveyResult(String animal, String breed, int age, int bcs, int weight, HttpServletRequest request, Model model) {
+	public String surveyResult(String animal, String breed, int age, int bcs, double weight, HttpServletRequest request, Model model) {
 
 		int userNo = 2; // session
